refactor: restaurant 페이지 컴포넌트를 RestaurantPage로 분리, SSR/CSR 전환 용이 #101
